src/docker/consumer/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
dataclasses==0.6
itsdangerous==2.0.1
jinja2==3.0.3
markupsafe==2.0.1
pip==21.2.4
python-dateutil==2.8.2
pytz==2021.3
setuptools==58.0.4
six==1.16.0
werkzeug==2.0.3
wheel==0.37.1
wincertstore==0.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
urllib3==1.26.7
requests==2.26.0 
paho-mqtt==1.6.1